Refactor WS API
